@@ -3,9 +3,11 @@ package ru.vadim.hostel.actors;
 import akka.actor.Props;
 import akka.actor.UntypedActor;
 import akka.japi.Creator;
+import org.springframework.context.annotation.Scope;
 import ru.vadim.hostel.entity.Category;
 import ru.vadim.hostel.repository.CategoryRepository;
 
+@Scope("prototype")
 public class CategoryActor extends UntypedActor {
 
     private CategoryRepository categoryRepository;
@@ -26,15 +28,13 @@ public class CategoryActor extends UntypedActor {
     }
 
     @Override
-    public void onReceive(Object message)  {
-        try {
-            if (message instanceof Category category) {
+    public void onReceive(Object message) throws Exception  {
+            if (message instanceof Category) {
+                Category category = (Category) message;
+                System.out.println(category);
                 categoryRepository.save(category);
             } else {
                 unhandled(message);
             }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
     }
 }
